Start processing I2C data after first byte
RPi2 stopped sending stop conditions between frames of an I2C request
(repeated start), which means this library does not detect the step
between sending data and expecting a processed result.

As we only ever send one byte of data, it is safe for now to process the
buffer after the first byte to have the result ready when it is expected.